:label: [rename] screen의 chat 폴더 추가
